# api/swagger.yaml
basePath: /beetle
definitions:
...
  controller.JSONResult:
    type: object
...
  model.BastionNode:
    properties:
      mciId:
        type: string
      vmId:
        type: string
    type: object
  model.ConnConfig:
    properties:
      configName:
        type: string
      credentialHolder:
        type: string
      credentialName:
        type: string
      driverName:
        type: string
      providerName:
        type: string
      regionDetail:
        $ref: '#/definitions/model.RegionDetail'
      regionRepresentative:
        type: boolean
      regionZoneInfo:
        $ref: '#/definitions/model.RegionZoneInfo'
      regionZoneInfoName:
        type: string
      verified:
        type: boolean
    type: object
  model.IdList:
    properties:
      output:
        items:
          type: string
        type: array
    type: object
...
  model.KeyValue:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
  model.Location:
    properties:
      display:
        type: string
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
  model.RegionDetail:
    properties:
      description:
        type: string
      location:
        $ref: '#/definitions/model.Location'
      regionId:
        type: string
      regionName:
        type: string
      zones:
        items:
          type: string
        type: array
    type: object
...
  model.RegionZoneInfo:
    properties:
      assignedRegion:
        type: string
      assignedZone:
        type: string
    type: object
  model.SimpleMsg:
    properties:
      message:
        example: Any message
        type: string
    type: object
...
  model.TbFirewallRuleInfo:
    properties:
      cidr:
        type: string
      direction:
        description: '`json:"direction"`'
        type: string
      fromPort:
        description: '`json:"fromPort"`'
        type: string
      ipprotocol:
        description: '`json:"ipProtocol"`'
        type: string
      toPort:
        description: '`json:"toPort"`'
        type: string
    required:
    - direction
    - fromPort
    - ipprotocol
    - toPort
    type: object
...
  model.TbSecurityGroupInfo:
    properties:
      associatedObjectList:
        items:
          type: string
        type: array
      connectionConfig:
        $ref: '#/definitions/model.ConnConfig'
      connectionName:
        type: string
      cspResourceId:
        description: CspResourceId is resource identifier managed by CSP
        example: csp-06eb41e14121c550a
        type: string
      cspResourceName:
        description: CspResourceName is name assigned to the CSP resource. This name
          is internally used to handle the resource.
        example: we12fawefadf1221edcf
        type: string
      description:
        type: string
      firewallRules:
        items:
          $ref: '#/definitions/model.TbFirewallRuleInfo'
        type: array
      id:
        description: Id is unique identifier for the object
        example: aws-ap-southeast-1
        type: string
      isAutoGenerated:
        type: boolean
      keyValueList:
        items:
          $ref: '#/definitions/model.KeyValue'
        type: array
      name:
        description: Name is human-readable string to represent the object
        example: aws-ap-southeast-1
        type: string
      resourceType:
        description: ResourceType is the type of the resource
        type: string
      systemLabel:
        description: SystemLabel is for describing the Resource in a keyword (any
          string can be used) for special System purpose
        example: Managed by CB-Tumblebug
        type: string
      uid:
        description: Uid is universally unique identifier for the object, used for
          labelSelector
        example: wef12awefadf1221edcf
        type: string
      vNetId:
        type: string
    type: object
  model.TbSecurityGroupReq:
    properties:
      connectionName:
        type: string
      cspResourceId:
        description: CspResourceId is required to register object from CSP (option=register)
        type: string
      description:
        type: string
      firewallRules:
        description: validate:"required"`
        items:
          $ref: '#/definitions/model.TbFirewallRuleInfo'
        type: array
      name:
        type: string
      vNetId:
        type: string
    required:
    - connectionName
    - name
    - vNetId
    type: object
...
  model.TbSubnetInfo:
    properties:
      bastionNodes:
        items:
          $ref: '#/definitions/model.BastionNode'
        type: array
      connectionConfig:
        $ref: '#/definitions/model.ConnConfig'
      connectionName:
        type: string
      cspResourceId:
        description: CspResourceId is resource identifier managed by CSP
        example: csp-06eb41e14121c550a
        type: string
      cspResourceName:
        description: CspResourceName is name assigned to the CSP resource. This name
          is internally used to handle the resource.
        example: we12fawefadf1221edcf
        type: string
      cspVNetId:
        description: CspVNetId is vNet resource identifier managed by CSP
        example: csp-45eb41e14121c550a
        type: string
      cspVNetName:
        description: CspVNetName is identifier to handle CSP vNet resource
        example: we12fawefadf1221edcf
        type: string
      description:
        type: string
      id:
        description: Id is unique identifier for the object
        example: aws-ap-southeast-1
        type: string
      ipv4_CIDR:
        type: string
      keyValueList:
        items:
          $ref: '#/definitions/model.KeyValue'
        type: array
      name:
        description: Name is human-readable string to represent the object
        example: aws-ap-southeast-1
        type: string
      resourceType:
        description: ResourceType is the type of the resource
        type: string
      status:
        type: string
      uid:
        description: Uid is universally unique identifier for the object, used for
          labelSelector
        example: wef12awefadf1221edcf
        type: string
      zone:
        type: string
    type: object
  model.TbSubnetReq:
    properties:
      description:
        example: subnet00 managed by CB-Tumblebug
        type: string
      ipv4_CIDR:
        example: 10.0.1.0/24
        type: string
      name:
        example: subnet00
        type: string
      zone:
        type: string
    required:
    - ipv4_CIDR
    - name
    type: object
  model.TbVNetInfo:
    properties:
      associatedObjectList:
        items:
          type: string
        type: array
      cidrBlock:
        type: string
      connectionConfig:
        $ref: '#/definitions/model.ConnConfig'
      connectionName:
        type: string
      cspResourceId:
        description: CspResourceId is resource identifier managed by CSP
        example: csp-06eb41e14121c550a
        type: string
      cspResourceName:
        description: CspResourceName is name assigned to the CSP resource. This name
          is internally used to handle the resource.
        example: we12fawefadf1221edcf
        type: string
      description:
        type: string
      id:
        description: Id is unique identifier for the object
        example: aws-ap-southeast-1
        type: string
      isAutoGenerated:
        type: boolean
      keyValueList:
        items:
          $ref: '#/definitions/model.KeyValue'
        type: array
      name:
        description: Name is human-readable string to represent the object
        example: aws-ap-southeast-1
        type: string
      resourceType:
        description: ResourceType is the type of the resource
        type: string
      status:
        type: string
      subnetInfoList:
        items:
          $ref: '#/definitions/model.TbSubnetInfo'
        type: array
      systemLabel:
        description: SystemLabel is for describing the Resource in a keyword (any
          string can be used) for special System purpose
        example: Managed by CB-Tumblebug
        type: string
      uid:
        description: Uid is universally unique identifier for the object, used for
          labelSelector
        example: wef12awefadf1221edcf
        type: string
    type: object
  model.TbVNetReq:
    properties:
      cidrBlock:
        example: 10.0.0.0/16
        type: string
      connectionName:
        example: aws-ap-northeast-2
        type: string
      description:
        example: vnet00 managed by CB-Tumblebug
        type: string
      name:
        example: vnet00
        type: string
      subnetInfoList:
        items:
          $ref: '#/definitions/model.TbSubnetReq'
        type: array
    required:
    - connectionName
    - name
    type: object
...
  resource.RestGetAllSecurityGroupResponse:
    properties:
      securityGroup:
        items:
          $ref: '#/definitions/model.TbSecurityGroupInfo'
        type: array
    type: object
  resource.RestGetAllVNetResponse:
    properties:
      vNet:
        items:
          $ref: '#/definitions/model.TbVNetInfo'
        type: array
    type: object
...
host: localhost:8056
info:
  contact:
    name: API Support
    url: https://github.com/cloud-barista/cm-beetle/issues/new/choose
  description: CM-Beetle REST API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: CM-Beetle REST API
  version: latest
paths:
...
  /migration/ns/{nsId}/resources/securityGroup:
    get:
      consumes:
      - application/json
      description: Get the list of all migrated security groups in the namespace
      operationId: ListMigratedSecurityGroups
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Option
        enum:
        - id
        in: query
        name: option
        type: string
      - description: 'Field key for filtering (ex: systemLabel)'
        in: query
        name: filterKey
        type: string
      - description: 'Field value for filtering (ex: Registered from CSP resource)'
        in: query
        name: filterVal
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Different return structures by the given option param
          schema:
            allOf:
            - $ref: '#/definitions/controller.JSONResult'
            - properties:
                '[DEFAULT]':
                  $ref: '#/definitions/resource.RestGetAllSecurityGroupResponse'
                '[ID]':
                  $ref: '#/definitions/model.IdList'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: List all migrated security groups
      tags:
      - '[Migration] Resources for VM infrastructure'
    post:
      consumes:
      - application/json
      description: Create a new migrated security group in the namespace
      operationId: CreateMigratedSecurityGroup
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: 'Option: [required params for register] connectionName, name,
          vNetId, cspResourceId'
        enum:
        - register
        in: query
        name: option
        type: string
      - description: Details for an securityGroup object
        in: body
        name: securityGroupReq
        required: true
        schema:
          $ref: '#/definitions/model.TbSecurityGroupReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbSecurityGroupInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Create a migrated security group
      tags:
      - '[Migration] Resources for VM infrastructure'
  /migration/ns/{nsId}/resources/securityGroup/{securityGroupId}:
    delete:
      consumes:
      - application/json
      description: Delete a specific migrated security group in the namespace
      operationId: DeleteMigratedSecurityGroup
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Security Group ID
        in: path
        name: securityGroupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Delete a migrated security group
      tags:
      - '[Migration] Resources for VM infrastructure'
  /migration/ns/{nsId}/resources/securityGroup/{sgId}:
    get:
      consumes:
      - application/json
      description: Get details of a specific migrated security group in the namespace
      operationId: GetMigratedSecurityGroup
      parameters:
      - description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Security Group ID
        in: path
        name: sgId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbSecurityGroupInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Get a specific migrated security group
      tags:
      - '[Migration] Resources for VM infrastructure'
  /migration/ns/{nsId}/resources/vNet:
    get:
      consumes:
      - application/json
      description: Get the list of all migrated virtual networks in the namespace
      operationId: ListMigratedVNet
      parameters:
      - description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Different return structures by the given option param
          schema:
            allOf:
            - $ref: '#/definitions/controller.JSONResult'
            - properties:
                '[DEFAULT]':
                  $ref: '#/definitions/resource.RestGetAllVNetResponse'
                '[ID]':
                  $ref: '#/definitions/model.IdList'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: List all migrated virtual networks
      tags:
      - '[Migration] Resources for VM infrastructure'
    post:
      consumes:
      - application/json
      description: Create a new migrated virtual network in the namespace
      operationId: CreateVNet
      parameters:
      - description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Virtual Network creation request
        in: body
        name: vNetReq
        required: true
        schema:
          $ref: '#/definitions/model.TbVNetReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbVNetInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Create a migrated virtual network
      tags:
      - '[Migration] Resources for VM infrastructure'
  /migration/ns/{nsId}/resources/vNet/{vNetId}:
    delete:
      consumes:
      - application/json
      description: Delete a specific migrated virtual network in the namespace
      operationId: DeleteMigratedVNet
      parameters:
      - description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Virtual Network ID
        in: path
        name: vNetId
        required: true
        type: string
      - description: Action
        enum:
        - withsubnets
        - refine
        - force
        in: query
        name: action
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Delete a migrated virtual network
      tags:
      - '[Migration] Resources for VM infrastructure'
    get:
      consumes:
      - application/json
      description: Get details of a specific virtual network in the namespace
      operationId: GetMigratedVNet
      parameters:
      - description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Virtual Network ID
        in: path
        name: vNetId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbVNetInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Get a specific migrated virtual network
      tags:
      - '[Migration] Resources for VM infrastructure'
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"